+++ src/CodeIndex/Program.cs
@@ -60,6 +60,7 @@ return args[0] switch
     "inspect" => QueryCommandRunner.RunInspect(args[1..], jsonOptions),
     "outline" => QueryCommandRunner.RunOutline(args[1..], jsonOptions),
     "status" => QueryCommandRunner.RunStatus(args[1..], jsonOptions, appVersion),
+    "validate" => QueryCommandRunner.RunValidate(args[1..], jsonOptions),
     "languages" => QueryCommandRunner.RunLanguages(args[1..], jsonOptions),
     "deps" => QueryCommandRunner.RunDeps(args[1..], jsonOptions),
     "index" => IndexCommandRunner.Run(args[1..], jsonOptions),